requirements.txt
Django==2.2.24
djangorestframework==3.12.4
drf-writable-nested==0.7.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
